app/cloud_function/requirements.txt
pyTelegramBotAPI
python-dotenv
ydb
akinator.py